use liquibase to manage database fixes #4
