+++ src/main/resources/application.properties
@@ -1,7 +1,10 @@
 spring.datasource.url=jdbc:h2:mem:crossride;DB_CLOSE_ON_EXIT=true
+
 # Enabling H2 Console
 spring.h2.console.enabled=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=validate
+spring.liquibase.change-log=classpath:/db/changelog/db.changelog-master.xml
+
 spring.data.rest.base-path=/api
 logging.level.root=INFO
 logging.level.org.hibernate.SQL=INFO
