+++ src/main/resources/db/changelog/db.changelog-master.xml
@@ -1,0 +1,5 @@
+<databaseChangeLog xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                        http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
+    <include file="db/changelog/db.changelog-1.0.xml" />
+</databaseChangeLog>
